Started implementing endpoint of VMs in CourseController

+++ resource_server/src/main/java/it/polito/ai/lab2/services/VmServiceImpl.java
@@ -10,12 +10,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional
 public class VmServiceImpl implements VmService {
 
     @Autowired
@@ -51,33 +53,20 @@ public class VmServiceImpl implements VmService {
     }
 
     @Override
-    public boolean createVm(Long vmModelId, VmDTO vmDTO, String courseId) {
-        VmModel vmModel = vmModelRepository.findById(vmModelId).orElseThrow(() -> new VmModelNotFoundException("VmModel " + vmModelId + " does not exist"));
+    public VmDTO createVm(Long teamId, VmDTO vmDTO) {
+        Team team = teamRepository.findById(teamId).orElseThrow(() -> new TeamNotFoundException("Team " + teamId + " does not exist"));
         Student owner = studentRepository.findById(SecurityContextHolder.getContext().getAuthentication().getName()).orElseThrow(() -> new StudentNotFoundException("Student " + SecurityContextHolder.getContext().getAuthentication().getName() + " does not exist"));
-        Course course = courseRepository.findByName(courseId).orElseThrow(() -> new CourseNotFoundException("Course " + courseId + " does not exist"));
 
-        if (!owner.getCourses().contains(course)) { //student not enrolled in course
-            throw new StudentNotInCourseException("Student " + owner.getId() + " is not enrolled in course " + courseId);
+        if (!owner.getCourses().contains(team.getCourse())) { //student not enrolled in course
+            throw new StudentNotInCourseException("Student " + owner.getId() + " is not enrolled in course " + team.getCourse().getAcronym());
         }
 
-        if (!course.isEnabled()) { //course not enabled
-            throw new CourseNotEnabledException("Course " + courseId + " is not enabled");
+        if (!team.getCourse().isEnabled()) { //course not enabled
+            throw new CourseNotEnabledException("Course " + team.getCourse().getAcronym() + " is not enabled");
         }
 
         if (owner.getTeams().isEmpty()) { //student in no teams
             throw new StudentNotInTeamException("Student " + owner.getId() + " does not belong to any team");
-        }
-
-        Team team = null;
-        for (Team t : owner.getTeams()) { //find the student's team of the selected course
-            if (t.getCourse().getName().equals(courseId)) {
-                team = t;
-                break;
-            }
-        }
-
-        if (team == null) {
-            throw new StudentNotInTeamOfCourseException("Student " + owner.getId() + "does not belong to a team of course " + courseId);
         }
 
         if (team.getVms().size() + 1 > team.getMaxTotalInstances()) {
@@ -107,7 +96,7 @@ public class VmServiceImpl implements VmService {
         }
 
         Vm vm = new Vm();
-        vm.setVmModel(vmModel);
+        vm.setVmModel(team.getCourse().getVmModel());
         vm.setActive(vmDTO.isActive());
         vm.setVCpu(vmDTO.getVCpu());
         vm.setRam(vmDTO.getRam());
@@ -116,26 +105,34 @@ public class VmServiceImpl implements VmService {
         vm.setTeam(team);
         vm.addOwner(owner);
         vmRepository.save(vm);
-        return true;
+        return modelMapper.map(vm, VmDTO.class);
     }
 
     @Override
-    public boolean deleteVm(Long vmId) {
+    public VmDTO deleteVm(Long vmId, Long teamId) {
         Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
+        }
 
         if (vm.getOwners().contains(studentRepository.getOne(SecurityContextHolder.getContext().getAuthentication().getName())) && !vm.isActive()) {
             vmRepository.delete(vm);
-            return true;
+            return modelMapper.map(vm, VmDTO.class);
         }
         throw new CannotDeleteVmException("VM " + vmId + " cannot be deleted. VM is still active or you are not one of the owners");
     }
 
     @Override
-    public boolean updateVmResources(Long vmId, int vCpu, int diskStorage, int ram) { //only owners
+    public VmDTO updateVmResources(Long vmId, Long teamId, int vCpu, int diskStorage, int ram) { //only owners, TODO: implementa check con il principal
         Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
 
         if (vm.isActive()) {
             throw new VmIsActiveException("Vm " + vmId + " is active and cannot be updated. Please turn it off");
+        }
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
         }
 
         Team team = vm.getTeam();
@@ -158,39 +155,73 @@ public class VmServiceImpl implements VmService {
         vm.setRam(ram);
         vm.setDiskStorage(diskStorage);
         vmRepository.save(vm);
-        return true;
+        return modelMapper.map(vm, VmDTO.class);
     }
 
     @Override
-    public boolean turnOnVm(Long vmId) { //only owners
+    public VmDTO turnOnVm(Long vmId, Long teamId) { //only owners
         Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
+        Team team = teamRepository.findById(teamId).orElseThrow(() -> new TeamNotFoundException("Team " + teamId + " does not exist"));
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
+        }
+
+        int numOfActiveVms = 0;
+
+        for (Vm v : team.getVms()) {
+            if (v.isActive()) {
+                numOfActiveVms += 1;
+            }
+        }
+
+        if(numOfActiveVms + 1 > team.getMaxActiveInstances()){
+            throw new MaxVmResourcesException("Cannot turn on the VM, too many active VMs");
+        }
+
         vm.setActive(true);
-        return true;
+        vmRepository.save(vm);
+        return modelMapper.map(vm, VmDTO.class);
     }
 
     @Override
-    public boolean turnOffVm(Long vmId) { //only owners
+    public VmDTO turnOffVm(Long vmId, Long teamId) { //only owners
         Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
+        teamRepository.findById(teamId).orElseThrow(() -> new TeamNotFoundException("Team " + teamId + " does not exist"));
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
+        }
+
         vm.setActive(false);
-        return true;
+        vmRepository.save(vm);
+        return modelMapper.map(vm, VmDTO.class);
     }
 
     @Override
-    public boolean addVmOwner(Long vmId, String studentId) { //only owners
+    public void addVmOwner(Long vmId, Long teamId, String studentId) { //only owners
         Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
         Student student = studentRepository.findById(studentId).orElseThrow(() -> new StudentNotFoundException("Student " + studentId + " does not exist"));
+        teamRepository.findById(teamId).orElseThrow(() -> new TeamNotFoundException("Team " + teamId + " does not exist"));
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
+        }
 
         if (!vm.getOwners().contains(student)) {
             vm.addOwner(student);
-            return true;
         }
-        return false;
     }
 
     @Override
-    public Optional<VmDTO> getVm(Long vmId) {
-        return vmRepository.findById(vmId)
-                .map(vm -> modelMapper.map(vm, VmDTO.class));
+    public VmDTO getVm(Long vmId, Long teamId) {
+        Vm vm = vmRepository.findById(vmId).orElseThrow(() -> new VmNotFoundException("Vm " + vmId + " does not exist"));
+
+        if(!vm.getTeam().getId().equals(teamId)){
+            throw new VmNotOfTeamException("Vm " + vmId + " does not belong to team " + teamId);
+        }
+
+        return modelMapper.map(vm, VmDTO.class);
     }
 
     @Override
